Move Vote subpages to their own files
